+++ Lecteur/lecteur.h
@@ -15,23 +15,23 @@
 
 class Lecteur {
 public:
-   Lecteur(int _id, std::string _name="Inconnu", std::string _surname="Inconnu", std::list<int> _liste={});
+   Lecteur(int _id, std::string _name="Inconnu", std::string _surname="Inconnu", std::list<long int> _liste={});
    std::string name() const;
    std::string surname() const;
    int id() const;
-   std::list<int> liste() const;
+   std::list<long int> liste() const;
    void updateId(int newId);
-   void addToList(int newISBN);
+   void addToList(long int newISBN);
    
 private:
    int _id;
    std::string _name;
    std::string _surname;
-   std::list<int> _liste;
+   std::list<long int> _liste;
    
 };
 
-std::ostream& operator<< (std::ostream& os, const std::list<int>& liste);
+std::ostream& operator<< (std::ostream& os, const std::list<long int>& liste);
 
 #endif // DATE_H
